requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Bottleneck==1.4.0
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.1.52
langchain-openai==0.0.8
langsmith==0.1.99
MarkupSafe==2.1.5
mmh3==4.1.0
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
openai==1.40.3
orjson==3.10.7
packaging==23.2
pinecone-client==3.1.0
pinecone-text==0.9.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.5
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.6.0
tqdm==4.66.5
types-requests==2.32.0.20240712
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wget==3.2
yarl==1.9.4
